working on user orders

+++ app/Livewire/Frontend/Checkout/Checkout.php
@@ -2,6 +2,7 @@
 
 namespace App\Livewire\Frontend\Checkout;
 
+use App\Helpers\NumberHelper;
 use App\Models\Cart;
 use App\Models\Order;
 use App\Models\OrderItem;
@@ -52,36 +53,45 @@ class Checkout extends Component
                     return $cartItem->product->price_per_kg * $cartItem->cart_count;
                 });
             $totalCharge = $shippingCharge + $subTotal;
-            $order = Order::create([
-                'user_id' => $user->id,
-                'sub_total' => $subTotal,
-                'shipping_charge' => $shippingCharge,
-                'total_charge' => $totalCharge,
-            ]);
-
-            // store order items in order items table
-            $carts = Cart::with('product')->where('user_id', Auth::user()->id)->get();
-            foreach ($carts as $key => $cart) {
-                OrderItem::create([
-                    'order_id' => $order->id,
-                    'product_id' => $cart->product_id,
-                    'product_name' => $cart->product->name,
-                    'qty' => $cart->cart_count,
-                    'price' => $cart->product->price_per_kg,
-                    'total' => $cart->cart_count * $cart->product->price_per_kg + $shippingCharge,
+            $userBalance = Auth::user()->account->balance;
+            if ($userBalance < $totalCharge) {
+                $userBalanceNepali = NumberHelper::toNepaliNumber($userBalance);
+                $this->dispatch('warningMessage', title: "तपाईंको खाता मा पर्याप्त रकम छैन। तपाईंको शेष रकम रु. $userBalanceNepali छ।");
+                return;
+            } else {
+                $order = Order::create([
+                    'user_id' => $user->id,
+                    'sub_total' => $subTotal,
+                    'shipping_charge' => $shippingCharge,
+                    'total_charge' => $totalCharge,
                 ]);
-                // $product=Product::where('id',$cart->product_id)->first();
-                // if($product->track_qty==true){
-                //     $product->update(
-                //         [
-                //             'qty'=>$product->qty-$cart->cart_count
-                //         ]
-                //     );
-                // }
+
+                // store order items in order items table
+                $carts = Cart::with('product')->where('user_id', Auth::user()->id)->get();
+                foreach ($carts as $key => $cart) {
+                    OrderItem::create([
+                        'order_id' => $order->id,
+                        'product_id' => $cart->product_id,
+                        'product_name' => $cart->product->name,
+                        'qty' => $cart->cart_count,
+                        'price' => $cart->product->price_per_kg,
+                        'total' => $cart->cart_count * $cart->product->price_per_kg + $shippingCharge,
+                    ]);
+                    // $product=Product::where('id',$cart->product_id)->first();
+                    // if($product->track_qty==true){
+                    //     $product->update(
+                    //         [
+                    //             'qty'=>$product->qty-$cart->cart_count
+                    //         ]
+                    //     );
+                    // }
+                }
+                Cart::where('user_id', Auth::user()->id)->delete();
+                // Redirect to checkout route
+                return redirect()->route('frontend.checkout.success');
             }
-            Cart::where('user_id', Auth::user()->id)->delete();
         } catch (\Throwable $th) {
-            $this->dispatch('error',$th->getMessage());
+            $this->dispatch('error', $th->getMessage());
         }
     }
 }
